Move field get (via handle) to top level class
